Add files via upload
Key Skills: C++ Language
The Project: “ONLINE VOTING SYSTEM” is an online voting technique. In this system people whose age is 18 years or above and of any sex can give his/her vote through online mode without going to the polling stations.
The main objective of this project is that we have to compare our online voting system data to other techniques of voting. And also, it will provide a means to conduct a less stressful and fair elections